fix: superclass's id should properly map to it's children

+++ src/main/java/com/museubd/model/bean/Outros.java
@@ -15,7 +15,7 @@ public class Outros implements Serializable {
 
     @Id
     @Column(name = "num_obj3")
-    private int num_obj3;
+    private Long num_obj3;
 
     @Column(name = "tipo")
     private String tipo;
